+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   Ganache based `fork_network`.
 - Move internal test suite to use Anvil instead of Ganache. This allows us to remove
   `flaky` decorators on tests.
+- Add `analysis.py` for Uniswap V3 and relevant tests
+- Add `mock_partial_deployment` function for V3
+- Abstract `TradeResult`, `TradeSuccess`, and `TradeFailure` out of Uniswap V2 and into eth_defi.trade, since also used in Uniswap V3
+
 
 # 0.13.11